feat: map_height 구함#18
리팩토링 완료, str이 NULL일 때와 개행일 때 종료하는 분기가 있는 gnl 실행이 있는데 이 부분에 str == NULL일 때 SIGSEGV가 발생하는 부분 해결
1.맵 이후에 있는 개행, 붙어있지 않은 맵, 유효하지 않은 문자 검사
2.map content가 정상이라면 map_height만큼 malloc한 뒤 gnl 반환값 넣기
3.가장 긴 문자열을 찾아 map_width에 넣고 height x width 크기의 배열 할당 받는다.
4.memset으로 공백 채우기
5.이후 기존 map 을 넣기
6.맵 구성 요소 검사
